requirements.txt
alembic==1.13.1
click==8.1.7
Faker==8.13.2
Flask==3.0.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
Jinja2==3.1.3
numpy==1.22.3
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
SQLAlchemy==2.0.25
